neural_devops_autopilot/requirements.txt
google-generativeai>=0.8.0
python-dotenv==1.0.1
streamlit==1.31.0
pandas==2.2.0